requirements.txt
amqp==5.2.0
asgiref==3.8.1
beautifulsoup4==4.12.3
billiard==4.2.1
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.16
django-bootstrap-v5==1.0.11
djangorestframework==3.15.2
kombu==5.4.2
Markdown==3.7
pillow==11.0.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.0
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
tzdata==2024.2
vine==5.1.0
wcwidth==0.2.13
